Updated openstack/openstack Project: openstack/nova  f587a78f4a1da017427a729334c69ec3bdcaf519
Cleanup of Translations

In preparation for the release, do some cleanups for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

This change needs to be done manually since the automatic import does
not handle some of these cases.

Change-Id: I4a2ed83862c7b32a0a93bfb1f040dfbe9e1d03af