Updated openstack/openstack Project: openstack/python-barbicanclient  5700a0fea3178879c3e657df6b8bdcb90544d8e7
Create Openstack CLI plugin for Barbican

This is the first implementation of the barbicanclient plugin for the
OpenStack CLI. Since we chose to use cliff in for the client, the
integration with the OCC is quite seamless. Only change that needed to
be done was that, since the OCC uses a ClientManager class to fetch the
specific libraries for the components, we needed to simulate that in the
barbicanclient.

Currently, the plugin can be used if the barbican server is configured
with keystone authentication. This is because currently there is no way
for us to set the X-Project-Id header manually from the client (if the
plugin is being used) but that functionality is out of the scope of this
first implementation.

Change-Id: I37fd158af24e785bc0b7125c6b4c1a9795927a10