Updated openstack/openstack Project: openstack/cinder  78cd21b1c9c429792024596f608c6a6d5d583fa2
Fix VMAX live migration problem

Live migration fails for VMAX because the host is not taken into
account when determining whether a volume is masked or not. For
live migration it is necessary to know which host a volume is masked
to. This patch fixes the problem by checking which host the volume
is masked to.

Change-Id: I693e29f7b26145bd1fd357b7d98377e26bfdfed8
Closes-Bug: #1498964