Updated openstack/openstack Project: openstack/neutron  e9e13fb19aa789b21f3b80cf60837cddc79cd1ee
usage_audit: Fix usage_audit to work with ML2

The script didn't work with the ML2 plugin as router related
functionality is moved to the L3 router service plugin. The fix assumes
that the L3 router service plugin is in use, and this seems to apply to
most of the current core plugins.

Change-Id: If3eb7bde4a5b8b9ca2ac0bd7327325f6c0ad620a
Closes-Bug: 1481692