Updated openstack/openstack Project: openstack/ironic  87381e2f47dd37d53343bd630d734a0155501f1a
Replace six.iteritems() with .items()

As mentioned in the developer's guide[0] and in a discussion[1],
we shouldn't use six.iteritems() unless there's a good reason (eg
large data set).

Since there's no need to use six.iteritems() in the two places
where it is being used, this patch replaces them with items().

[0] http://docs.openstack.org/infra/manual/developers.html#peer-review
[1] http://lists.openstack.org/pipermail/openstack-dev/2015-June/066391.html

Change-Id: I8c659550d8e70c217b43a4e64b178f1a91d2be77