Updated openstack/openstack Project: openstack-infra/devstack-gate  1f59e010ac41a504fa6797e62b6ef1bf42b4d8ea
Setup multinode interfaces on singlenode n-net

Because nova net takes over eth0 by default and slurps it into br100
there are corner cases where this will break funny existing network
setups. One example is where the VMs default route is not within the
network assigned to eth0.

Work around this by using the overlay networking with nova net always.
We will create an ovs gre bridge that is a bridge to nowhere. But nova
net can connect to that and not break the existing networking.

Change-Id: Ia8af343b17e05191e47dc7f3969aeaf19a443da6