Updated openstack/openstack Project: openstack/neutron-vpnaas  77e6a5e2e582ac8f73fcbaf2aba5587f0cb4d835
Tag the alembic migration revisions for Liberty

Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.

So from now on we tag the final revision on every branch when we make
a milestone release.

Partial-Bug: #1499033

Depends-On: I38623986dd574bec01fe147f9c6a747f3f512bb7

Change-Id: I237cd984273212da930794414c761f7748e6bd0c