Updated openstack/openstack Project: openstack/heat  87f5130e2c21c3ff5fe2482a03143307207fd2c1
Reduce frequency of logging environment resources

Currently we log all template resources every time any API call
happens which indirectly creates an Environment object, which
results in a lot of duplicate bloat in the heat logs (not even
at debug level..)

So, instead log all environment resources (e.g those from resource
plugins and the global environment) only when we start the
service, and log user-provided template resources only when we
do the template validation on create/update, for only the root stack,
because we pass a derived subset environment down to all children.

We also switch to using the string representation of the ResourceInfo
objects, as this contains a bit more useful info than the current
log format.

Change-Id: I20007c1ad6bf8b1479c1d60a5c392b8b9e1c06c4
Closes-Bug: #1499330