Updated openstack/openstack Project: openstack-infra/tripleo-ci  ba17c7a410f0e66c9587647c2691c6f618d5f3d1
Use tripleo.sh to drive ci

tripleo.sh is a script to automate a TripleO setup, holding as much of the
deploy logic in this script means it can be shared with developers. At the
moment the script holds most of the same logic as we have in ci but it will
develop over time.

Change-Id: Ie425cd59bb02d50688855c2b2c3304fa1f17489d
Depends-on: I9396a487964955f59584965ecbfd4c88cca87c0e