Updated openstack/openstack Project: openstack/keystone  dd9459fcaf7641736ec715fce0dbac6ecf9176d9
Update bandit blacklist_calls config

Bandit 0.13.2 was released and this is the min version in
requirements.txt. This version updated the standard
configuration for the blacklist_calls test.

With the new config, use of random.random is flagged by bandit
to indicate that developers need to verify that this isn't
used in secure context. The use in keystone is not for
security, so it's marked in the code as not an issue.

Change-Id: Icb72688df294afa13ac1643e26748c07cc8f0109