Updated openstack/openstack Project: openstack/horizon  ae13e41cf14b5159d07bca81c41d8ae7c31deacd
Enable pagination for Project->Images view

This feature is already implemented, but not enabled because it
has strange behavior when the image list is filtered by removing aki
and ari type of images.

The pagination for images view should be enabled since the filtering
image code was removed in: https://review.openstack.org/#/c/59935/

Because the ImagesTable is the only one being incorporated into the
images view, there is no need to include table.name attribute into
_prev and _more attributes of the view (no other tables to distinguish
between).

This is a refreshment of https://review.openstack.org/#/c/58928/
(which I cannot update because it's abandoned).

Change-Id: Ib027a548ee8aa623f0faea86f0ad7570d6ec2360
Co-Authored-By: chenhaiq <chenhaiq@cn.ibm.com>
Closes-Bug: #1252649