Updated openstack/openstack Project: openstack/neutron-fwaas  8f29c3a6ba4359b427ccc060861c29fe08cf438a
Tag the alembic migration revisions for Liberty

Previously when we had one repo with one alembic branch we would create a
milestone revision on that single branch. Now we have multiple repos and
expand/contract branches for each repo.

So from now on we tag the final revision on every branch when we make a
milestone release.

Partial-Bug: #1499033

Change-Id: I76566e33030b08479446e070f49472caa24fc600