Updated openstack/openstack Project: openstack/horizon  a23a7e0b6cbdafd2ce27a5f128e63c3ec0be5348
Show external networks

Only shared networks were considered as external.
The Neutron policy since Icehouse is router:external=True and shared=False so
external networks were not seen anymore
We need to be able to see the external network to obtain its ID (e.g. for
heat templates in the orchestration panel).

We now display external networks and added an "external" field in the networks
view

Change-Id: Ice1f4e2c11825b8d86a8a126350e71a88741acb7
Closes-Bug: #1472281