Updated openstack/openstack Project: openstack-infra/project-config  1d4a0b04138656ccfc2770249eab68d92df44954
Setup periodic stable liberty jobs

With the recent creation of the liberty branches we need to add the
periodic stable jobs to sanity check the state of the stable branches
to see if there are any issues with the branch. This commit adds the
necessary bit to start running these jobs.

Change-Id: I9285a2a1b0054b4d360175ced1a3e035be7fc1b6