Updated openstack/openstack Project: openstack/cinder  3c367411326a509d3764507bd42834cc7872a4ea
VMware: Remove VMDK driver for ESX server

The VMDK driver for ESX server was deprecated during
Icehouse release. Configuring Cinder to use this driver
prints a warning that the driver is deprecated and
will be removed in the Juno release.

This patch removes the deprecated VMDK driver for
ESX server.

Closes-Bug: #1494701
Change-Id: I8c0ad19cf09016c112d242f93b24c6742acaa1e8