Updated openstack/openstack Project: openstack/python-ironicclient  90ea9068bb6ad3122d385c6a617e2d934d5d7296
Replace six.iteritems() with .items()

As mentioned in the developer's guide[0] and in a discussion[1],
we shouldn't use six.iteritems() unless there's a good reason (eg
large data set).

Since there's no need to use six.iteritems() in the places where
it is being used, this patch replaces them with items().

[0] http://docs.openstack.org/infra/manual/developers.html#peer-review
[1] http://lists.openstack.org/pipermail/openstack-dev/2015-June/066391.html

Change-Id: If91afd37a7bf42dcbcf9dcd70fbbe82787de850e