Updated openstack/openstack Project: openstack/neutron  4071c14b10854e0b589bb9e5eb602ddcfc58a1d8
Remove is_ha property from the router

This was a todo left over from refactoring.  The one place where it
was used was in the metadata driver which can just as easily check
isinstance.

The is_ha property in the HA class only returned True if self.router
was True.  The agent sets this on creation and so it will always be
True.  Additionally, it didn't really make sense to me anyway given
the way this was being used in the metadata driver.

Change-Id: Idf91ee5f8b72ff9a1740dd60090327d07ee8e16e