Updated openstack/openstack Project: openstack-infra/groups-static-pages  39163864ce11a7e61d9467a3c221629409d589f8
Only need one URL for Official User Group

We don't want to unnecssarily encourage additional URLs for the sake of URLs.
So clarify that only one URL is needed to pass the step.

Also naturalise some of the language and update it to reflect the
implementation in the groups portal.

Change-Id: Icd1559eadd270a82ea53fcce8e58890741571a8d