Updated openstack/openstack Project: openstack/ironic-specs  529d9f76311ea1271c90a790667acdd56da4b649
Reorganize the specs on main page

This reorganizes the specs on the main page
(http://specs.openstack.org/openstack/ironic-specs/)
so that it is clearer as to which specs have been
approved (but not implemented) and which specs
were implemented (and in which development cycle).

Since specs that were approved, but not implemented in the same
cycle as the approval, do not have to be re-submitted, there
doesn't seem to be much value to having 'Current Cycle' as a
header.

Change-Id: I71e4fa1e60cfed114a8cc051a774a9d148fbbd8a