Updated openstack/openstack Project: openstack-infra/project-config  8e5b0405a4056b39d51301ce3719964c39c6e167
Enable ceilometer devstack plugin for tempest jobs

With the removal of ceilometer from devstack we need to update the
way in which tempest jobs are started for ceilometer: New jobs are used
to start tempest tests with the right backends and with the devstack
plugin in place. The 'integrated-gate' jobs are no longer used as
those do not start ceilometer so are not useful.

With this change we add the following jobs:

 gate-tempest-dsvm-ceilometer-mysql-neutron-full
 gate-tempest-dsvm-ceilometer-mysql-full
 gate-tempest-dsvm-ceilometer-postgresql-full

which effectively replace:

 gate-tempest-dsvm-full
 gate-tempest-dsvm-postgres-full
 gate-tempest-dsvm-neutron-full

The jobs are set up so that future expansion should not be too
terribly difficult.

Note: These have not been added as experimental because
ceilometer currently has no tempest coverage in the gate.

Change-Id: Ic74cc07462dfff1b4d55d837ea7218e7f3a4f1a3
Related-Bug: #1501739