Updated openstack/openstack Project: openstack-infra/puppet-cgit  1d272b55040cf806cf92058a4c8a6af9936311c0
Fix git-daemon refresh logic

Without this patch, the git-daemon init script file resource subscribes
to the systemd git-daemon socket file resource, and the exec that adds
or updates the git-daemon selinux port subscribes to changes in the
init script. The logic is broken here because a file resource cannot
subscribe to anything, only services and execs can subscribe to an
event. If the selinux exec needs to run again, for instance because the
git-daemon port has changed, it must wait for a change in the init
script. Since the init script is built from a static file and not a
template, it won't change if the git-daemon port changes, so the
selinux exec will not run.

This patch adds another subscribe relationship to the git-daemon
selinux exec on the git-daemon socket because if that changes, the
git-daemon exec needs to run again. We also replace the subscribe in
the init script resource with a require, which is a no-op change but
makes the relationship more clear.

Change-Id: Ia985dad81a95130ea55bb6479632375ac4ea6d24