Updated openstack/openstack Project: openstack/ironic  45127054b0a1ad9cd88cce0b8e5ce2147af078c0
Python 3.4 unit tests fail with LANG=C

After the move to os-testr the 'tox -epy34' unit tests will fail when
LANG=C is set in the environment.

Set LC_ALL=en_US.UTF-8 in the testenv to solve the issue. This is
in-line with our curent method of running the unit tests in US English
so that logging, messages, and exceptions can be validated by unit
tests.

Closes-Bug: #1501415
Change-Id: Ifefdd67c9e5edf22f3b7a1fce56f21c788ac9ead