Updated openstack/openstack Project: openstack/tempest  7d8be7702867e04c10f2614734d70369c803a721
Pass mount_path through to ssh_client.mount in scenario manager

The RemoteClient.mount method takes an optional mount_path kwarg which
defaults to '/mnt'. The create_timestamp and get_timestamp methods in
the base scenario manager also have the same kwarg but aren't passing it
though to the mount() method, which will break down if the caller in the
scenario tests ever uses a different mount path.

Closes-Bug: #1501512

Change-Id: I594f769178dd8a99398be2d590b8244517920a4c