Updated openstack/openstack Project: openstack/heat-translator  e6ba8b486a04240cdbe04df081ca7fc409db5d6d
Handle URLs for input templates and imports

Allow input templates and imported custom types to be provided as URLs
with Heat-Translator auto-detecting the type (file vs URL), and add
necessary unit tests.

Note, since for some test cases currently the required file does not
exist on github, we temporarily use files hosted somewhere else. Once
the patch is merged a follow-on patch will be submitted to fix that
issue and use URLs of the new files that are submitted with this patch.

Change-Id: I79c07c228d9d4de22f84ef384d46e3605b9a649e
Closes-Bug: #1340748
Partially Implements: blueprint tosca-namespaces