Updated openstack/openstack Project: openstack/keystone  c0ef67f84c72d1f140e0c0296906223ef39c7126
Update bandit blacklist_imports config

Bandit 0.13.2 was released and this is the min version in
requirements.txt. This version updated the standard
configuration for the blacklist_imports test.

With the new config, use of subprocess is flagged to indicate
that developers need to verify that this isn't used in secure
context. Since keystone uses subprocess in a special way due to
eventlet support this test is now disabled.

Change-Id: I020d3fa4fac4086b3264114a661326644e045df7