Updated openstack/openstack Project: openstack/tempest-lib  6d0b712e00dab4ad66bf79774f64c14b2434752c
Provide alternative ssh exec_command for non-linux environments

When running tests that use remote validation on OSX (or any
non-Linux OS), the exec_command method would raise an exception
due to the "poll" method not existing for select. While this
implementation is very safe and reliable, paramiko does provide
other means for handling reading the output from commands. This
patch provide a basic alternative to the existing method that is
only used when the poll method is not available.

Change-Id: Ic14b2331fd59a6311ebe51f598d2cc2320c4389e
Closes-Bug: #1480983