Updated openstack/openstack Project: openstack/manila  d5e5ee2404d44b0b25bb2e61cc61f839eab58596
Improve logging of calls in ShareManager

Consistently log main operations in the Share manager. Logs should
contain the status (success/error) and the share instance id.

This should help operators to know what was called and when (for debugging,
post mortem etc.).

We do that in the share manager so that every driver benefit from it.

Change-Id: Iad2653a2da851e35252a86ac6b10574969f0e409