Updated openstack/openstack Project: openstack/pycadf  9c4100b85d277b558143b1af9dc9ffdc37802e37
Add docstring validation

This introduces a linter for PEP257 to avoid trivial nitpicking of
docstrings in code reviews. Because flake8_docstrings simply provides a
plugin to add pep257 to flake8, you can run it via `tox -e pep8`.

PEP257 checks which we are currently violating are ignored in tox.ini.
We can remove them from the ignored list as they are fixed.

Related-Bug: 1501544
Change-Id: I01ebad7b70cf61dd80d3c06c6808d8178fbdd634